api/ocr/requirements.txt
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
isodate==0.7.2
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
